refactor: 패키지 구조 domain별로 분리하도록 리팩토링.JPAB-14-Refactor
