+++ src/main/java/com/prgrms/be/app/common/dto/PageRequest.java
@@ -1,37 +1,37 @@
-package com.prgrms.be.app.domain.dto;
+package com.prgrms.be.app.common.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
+
+import static org.springframework.data.domain.Sort.Direction;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostPageRequest {
+public class PageRequest {
 
     private static final int DEFAULT_SIZE = 10;
     private static final int MAX_SIZE = 20;
 
     private int page;
     private int size;
-    private Sort.Direction direction;
+    private Direction direction;
 
-    public PostPageRequest(int page, int size, Sort.Direction direction) {
+    public PageRequest(int page, int size) {
+        this(page, size, Direction.DESC);
+    }
+
+    public PageRequest(int page, int size, Direction direction) {
         this.page = page <= 0 ? 1 : page;
         this.size = isRangeSize(size) ? size : DEFAULT_SIZE;
         this.direction = direction;
-    }
-
-    public PostPageRequest(int page, int size) {
-        this(page, size, Sort.Direction.DESC);
     }
 
     private boolean isRangeSize(int size) {
         return size > 0 && size <= MAX_SIZE;
     }
 
-    public PageRequest of() {
-        return PageRequest.of(page - 1, size, direction, "createdAt");
+    public org.springframework.data.domain.PageRequest of() {
+        return org.springframework.data.domain.PageRequest.of(page - 1, size, direction, "createdAt");
     }
 }
