+++ src/main/java/com/prgrms/be/app/domain/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.prgrms.be.app.config;
+package com.prgrms.be.app.domain.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
