+++ src/main/java/com/prgrms/be/app/domain/post/dto/PostDetailResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.be.app.domain.dto;
+package com.prgrms.be.app.domain.post.dto;
 
 import java.time.LocalDateTime;
