+++ src/main/java/com/prgrms/be/app/domain/post/dto/PostsResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.be.app.domain.dto;
+package com.prgrms.be.app.domain.post.dto;
 
 import java.util.List;
