+++ src/main/java/com/prgrms/be/app/domain/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.prgrms.be.app.repository;
+package com.prgrms.be.app.domain.post.repository;
 
-import com.prgrms.be.app.domain.Post;
+import com.prgrms.be.app.domain.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
