+++ src/main/java/com/prgrms/be/app/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.prgrms.be.app.repository;
+package com.prgrms.be.app.domain.user.repository;
 
-import com.prgrms.be.app.domain.User;
+import com.prgrms.be.app.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
