+++ src/test/java/com/prgrms/be/app/domain/dto/PostDTOTest.java
@@ -1,5 +1,7 @@
 package com.prgrms.be.app.domain.dto;
 
+import com.prgrms.be.app.domain.post.dto.PostCreateRequest;
+import com.prgrms.be.app.domain.post.dto.PostUpdateRequest;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
